Refactor Verify page: moved verification logic into reusable VerifyForm component
